% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salbm_package.R
\docType{package}
\name{salbm-package}
\alias{salbm-package}
\alias{salbm}
\alias{##}
\alias{@useDynLib}
\alias{salbm,}
\alias{.registration}
\alias{=}
\alias{TRUE}
\title{The 'salbm' package.}
\description{
An analysis toolbox for missing data analysis
}
