% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salbmShiny.R
\name{salbmShiny}
\alias{salbmShiny}
\title{Run Web-Based \code{salbm} application}
\usage{
salbmShiny()
}
\description{
Call Shiny to run \code{salbm} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI of \code{salbm}.
}
\examples{
\dontrun{
salbmShiny()}
}
